reports

+++ Firmware/Board/v3/Inc/ffb.h
@@ -62,6 +62,6 @@ typedef struct EffectCalcData {
 void FFB_OnCreateNewEffect(const PID_CreateNewEffectReport* data);
 volatile const PID_BlockLoadReport *FFB_GetPidBlockLoad(void);
 
-void FFB_OnUsbData(uint8_t *buf, uint8_t len);
+void FFB_OnUsbData(uint8_t *buf, uint16_t len);
 
 #endif
